Move all wallet components into separate directory
