+++ src/components/wallet/get-wallet-error-message.tsx
@@ -8,7 +8,7 @@ import React from 'react';
 
 export const NO_ERROR = Symbol();
 
-export function getErrorMessage(err: unknown, fallbackMessage: React.ReactNode): React.ReactNode {
+export function getWalletErrorMessage(err: unknown, fallbackMessage: React.ReactNode): React.ReactNode {
     if (isWalletStandardError(err, WALLET_STANDARD_ERROR__FEATURES__WALLET_ACCOUNT_FEATURE_UNIMPLEMENTED)) {
         return (
             <>
