+++ src/components/wallet/unconnectable-wallet-menu-item.tsx
@@ -4,8 +4,8 @@ import { IconExclamationmarkTriangle } from 'symbols-react';
 import type { UiWallet } from '@wallet-standard/react';
 import { useState } from 'react';
 
-import { ErrorDialog } from './ErrorDialog';
-import { WalletMenuItemContent } from './WalletMenuItemContent';
+import { WalletErrorDialog } from './wallet-error-dialog';
+import { WalletMenuItemContent } from './wallet-menu-item-content';
 import {
     DropdownMenuItem,
 } from '@/components/ui/dropdown-menu';
@@ -32,7 +32,7 @@ export function UnconnectableWalletMenuItem({ error, wallet }: Props) {
             </DropdownMenuItem>
 
             {dialogIsOpen && (
-                <ErrorDialog 
+                <WalletErrorDialog
                     error={error} 
                     onClose={() => setDialogIsOpen(false)} 
                     title="Unconnectable wallet"
